% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{append.countrypair.dataframe}
\alias{append.countrypair.dataframe}
\title{This function appends to (Date, CountryPair) format data frame
data from (Date, Country) format}
\usage{
append.countrypair.dataframe(countrypair_df, country_df)
}
\description{
@import dplyr
}
