% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{calculate.gdp.df}
\alias{calculate.gdp.df}
\title{This function calculates absolute GDP values given growth rates
The growth rates are at quarterly frequency and the gdp balance is at}
\usage{
calculate.gdp.df(growth_rates, gdp_balance)
}
\description{
This function calculates absolute GDP values given growth rates
The growth rates are at quarterly frequency and the gdp balance is at
}
