% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{classify_crises_dates}
\alias{classify_crises_dates}
\title{This function classifies for given country the crises dates}
\usage{
classify_crises_dates(Target_Country, dates_vec, crises_df)
}
\description{
This function classifies for given country the crises dates
}
