% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{collapse_pair_controls}
\alias{collapse_pair_controls}
\title{This function collapses (calculates sum, difference and min)}
\usage{
collapse_pair_controls(fin_reg_df, control_vars,
  collapse_funcs = c("sum", "diff", "min"))
}
\description{
This function collapses (calculates sum, difference and min)
}
