% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{construct_countrypair_EU_index}
\alias{construct_countrypair_EU_index}
\title{This function constructs country pair EU index}
\usage{
construct_countrypair_EU_index(
  eu_df,
  dates_vec = NULL,
  countries = NULL,
  index_status = "both"
)
}
\description{
@import dplyr
}
