% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{construct_countrypair_harmon_index}
\alias{construct_countrypair_harmon_index}
\title{This function constructs country pair harmon index}
\usage{
construct_countrypair_harmon_index(df, dates_vec = NULL,
  index_status = "both")
}
\description{
The function takes a data frame with "Directive","Country","Date"
columns and calculates for each Date, Directive,CountryPair
the transposed status (0,1)
}
\details{
@import dplyr
}
