% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{deflate.data}
\alias{deflate.data}
\title{This helper function deflates bis data by US CPI}
\usage{
deflate.data(df, vars_to_deflate, cpi = NULL, remove_cpi_col = TRUE)
}
\description{
This helper function deflates bis data by US CPI
}
