% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{extract_id_from_imf_trade_filepath}
\alias{extract_id_from_imf_trade_filepath}
\title{Extract country and category form imf data filepath}
\usage{
extract_id_from_imf_trade_filepath(filepath)
}
\arguments{
\item{filepath}{string}
}
\description{
Extract country and category form imf data filepath
}
