% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_cycle.R
\name{get.clean.cycle}
\alias{get.clean.cycle}
\title{This function returns the cyclical component of time series
in case of insufficient observation the function returns NA}
\usage{
get.clean.cycle(timeseries, my_pl = 8, my_pu = 20, filter = "bk")
}
\description{
This function returns the cyclical component of time series
in case of insufficient observation the function returns NA
}
