% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.neg.abs.diff}
\alias{get.neg.abs.diff}
\title{This function calculates the negative absolute difference measure
The function takes a data frame with
Date, Country, VarName columns}
\usage{
get.neg.abs.diff(df)
}
\description{
This function calculates the negative absolute difference measure
The function takes a data frame with
Date, Country, VarName columns
}
