% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.panel.resid}
\alias{get.panel.resid}
\title{This function calculates the residuals of time and country fixed effect
The function takes a data frame with
Date, Country, VarName columns}
\usage{
get.panel.resid(df)
}
\description{
This function calculates the residuals of time and country fixed effect
The function takes a data frame with
Date, Country, VarName columns
}
