% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.quasi.cor}
\alias{get.quasi.cor}
\title{This function calculates the quasi correlation measure
The function takes a data frame with
Date, Country, VarName columns}
\usage{
get.quasi.cor(df)
}
\description{
This function calculates the quasi correlation measure
The function takes a data frame with
Date, Country, VarName columns
}
