% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{get.reg.list}
\alias{get.reg.list}
\title{This function performs regression alternating between "pop" and "gdp"
specs of explanatory vars}
\usage{
get.reg.list(reg_df, dep_var_name, xvars = NULL)
}
\description{
This function performs regression alternating between "pop" and "gdp"
specs of explanatory vars
}
