% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.roll.cor}
\alias{get.roll.cor}
\title{This function calculates the rolling pairwise correlation
The function takes a data frame with
Date, Country, VarName columns}
\usage{
get.roll.cor(df, win_len)
}
\description{
This function calculates the rolling pairwise correlation
The function takes a data frame with
Date, Country, VarName columns
}
