% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{get.sens.reg}
\alias{get.sens.reg}
\title{This function performs regressions on every dep var in the data frame}
\usage{
get.sens.reg(reg_df)
}
\description{
This function performs regressions on every dep var in the data frame
}
