% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{get.suffix.diff}
\alias{get.suffix.diff}
\title{This function calculates the difference of columns
with given suffixes}
\usage{
get.suffix.diff(df, suffix_A = "_A", suffix_B = "_B")
}
\description{
This function calculates the difference of columns
with given suffixes
}
