% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.synch2.list}
\alias{get.synch2.list}
\title{This function returns the synch2 measure
The function takes a data frame with
Date, Country, VarName(s) columns}
\usage{
get.synch2.list(df)
}
\description{
This function returns the synch2 measure
The function takes a data frame with
Date, Country, VarName(s) columns
}
