% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synch_measures_functions.R
\name{get.synch3.list}
\alias{get.synch3.list}
\title{This function returns the synch3 measure
The function takes a data frame with
Date, Country, VarName(s) columns}
\usage{
get.synch3.list(df, win_len)
}
\description{
This function returns the synch3 measure
The function takes a data frame with
Date, Country, VarName(s) columns
}
