% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{get_significant_names}
\alias{get_significant_names}
\title{This function extracts significant coefficient names from lm model}
\usage{
get_significant_names(temp_lm)
}
\description{
This function extracts significant coefficient names from lm model
}
