% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.bis.lbs.data}
\alias{import.bis.lbs.data}
\title{#' This function imports BIS LBS from scratch
#'
#' @import dplyr
#'
#' @import readr}
\usage{
import.bis.lbs.data(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\Documents\\\\Data\\\\BIS\\\\", "WEBSTATS_LBS_D_PUB_DATAFLOW_csv_col.csv"),
  my_instruments = "All instruments",
  my_measure = "Amounts outstanding / Stocks",
  my_currency = "All currencies",
  my_report_currency = "All currencies (=D+F+U)",
  my_lending_position = "Cross-border",
  my_reporting_institutions = paste0("All reporting", " banks/institutions ",
    "(domestic, foreign, consortium and ", "unclassified)"),
  my_counter_sector = "All sectors",
  countries_vec = NULL,
  collapse_countrypair = TRUE
)
}
\description{
#' This function imports BIS LBS from scratch
#'
#' @import dplyr
#'
#' @import readr
}
