% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.bis.property.price.data}
\alias{import.bis.property.price.data}
\title{This function imports BIS house price data from scratch}
\usage{
import.bis.property.price.data(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\Documents\\\\Data\\\\BIS\\\\", "WEBSTATS_SELECTED_PP_DATAFLOW_csv_col.csv"),
  my_value = "Real",
  my_measure = "Index, 2010 = 100",
  countries_vec = NULL
)
}
\description{
This function imports BIS house price data from scratch
}
