% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.bis.tot.credit.data}
\alias{import.bis.tot.credit.data}
\title{This function imports BIS total credit data from scratch}
\usage{
import.bis.tot.credit.data(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data\\\\BIS\\\\",
    "WEBSTATS_TOTAL_CREDIT_DATAFLOW_csv_col.csv"),
  my_lending_sector = "All sectors",
  my_unit_type = "US Dollar",
  my_valuation = "Market value",
  my_adjustment = "Adjusted for breaks",
  my_sector = "Private non-financial sector",
  countries_vec = NULL
)
}
\description{
This function imports BIS total credit data from scratch
}
