% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.crises.dates.df}
\alias{import.crises.dates.df}
\title{This helper function imports banking crises dates data}
\usage{

  import.crises.dates.df(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data\\\\LavaenValencia\\\\SYSTEMIC BANKING ",
  "CRISES DATABASE_2018.xlsx"), countries_vec = NULL)
}
\description{
This helper function imports banking crises dates data
}
