% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.fin.dev.ind}
\alias{import.fin.dev.ind}
\title{This function imports financial development data}
\usage{
import.fin.dev.ind(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\Documents\\\\Data\\\\Svirydzenka\\\\FinDev.xlsx"),
  countries_vec = NULL
)
}
\description{
@import readxl
}
\details{
@import dplyr
}
