% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.geodist.data}
\alias{import.geodist.data}
\title{This function imports geo dist data from cepii}
\usage{

  import.geodist.data(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents",
  "\\\\Data\\\\CEPII\\\\", "dist_cepii.xls"))
}
\description{
This function imports geo dist data from cepii
}
