% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.harmon.data}
\alias{import.harmon.data}
\title{This function imports harmon data}
\usage{
import.harmon.data(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data\\\\",
    "Kalemli_Ozcan_Papaionnou_Peydro\\\\harmon.xlsx"),
  myrange = "B4:AF32",
  codes_filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data\\\\ISO\\\\",
    "iso_2digit_alpha_country_codes.csv"),
  convert_country_names = TRUE
)
}
\description{
@import dplyr
}
\details{
@import readxl

 @import zoo
}
