% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_IMF_data.R
\name{import.imf.trade.data}
\alias{import.imf.trade.data}
\title{This function imports export import data from IMF Direction of
Trade data base}
\usage{
import.imf.trade.data(export_dirpath, import_dirpath)
}
\arguments{
\item{export_dirpath}{(optional) path to directory holding export
data files}

\item{import_dirpath}{(optional) path to directory holding import
data files}
}
\value{
data frame
}
\description{
This function imports export import data from IMF Direction of
Trade data base
}
