% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.iso.codes}
\alias{import.iso.codes}
\title{This function imports ISO codes for country names}
\usage{
import.iso.codes(filepath = NULL, type = "3-digits")
}
\description{
This function imports ISO codes for country names
}
