% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.kaopen.ind}
\alias{import.kaopen.ind}
\title{This function imports capital account openess (Chinn Ito) data}
\usage{

  import.kaopen.ind(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data\\\\Chin-Ito\\\\kaopen_2016.xls"),
  countries_vec = NULL)
}
\description{
@import readxl
}
\details{
@import dplyr
}
