% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.macropru.ind}
\alias{import.macropru.ind}
\title{This function imports macroprudential (Cerutti) data}
\usage{
import.macropru.ind(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"), "\\\\Documents",
    "\\\\Data\\\\Cerutti\\\\prudential_ind_3.xlsx"),
  countries_vec = NULL
)
}
\description{
@import readxl
}
\details{
@import dplyr
}
