% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_OECD_data.R
\name{import.oecd.gdp.quarter}
\alias{import.oecd.gdp.quarter}
\title{Import OECD GDP quarterly dataframe}
\usage{
import.oecd.gdp.quarter(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data", "\\\\OECD\\\\GDP_quarterly.csv"),
  my_subject = "TOT",
  my_measure = "PC_CHGPP",
  my_frequency = "Q"
)
}
\arguments{
\item{my_subject}{filters the required data type
(total - TOT or volume index - VOLIDX).}

\item{my_measure}{filters the required measurement type
(percentage change from the previous quarter  - PC_CHGPP,
 percentage change from the same quarter of the previous year - PC_CHGPY
 and volume index - IDX.) The default is PC_CHGPP.}

\item{my_frequency}{filters the required frequency
(annual - A or quarter - Q). The default is Q}
}
\description{
This function imports quarterly GDP data from OECD format.
}
\details{
This indicator is based on real GDP (also called GDP at constant prices
or GDP in volume), i.e. the developments over time are adjusted for price
changes. The numbers are also adjusted for seasonal influences.
The indicator is available in different measures: percentage change from
the previous quarter, percentage change from the same quarter of the
previous year and volume index. All OECD countries compile their data
according to the 2008 System of National Accounts (SNA).
}
