% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_OECD_data.R
\name{import.oecd.gdp}
\alias{import.oecd.gdp}
\title{Import OECD GDP dataframe}
\usage{
import.oecd.gdp(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data\\\\OECD\\\\GDP.csv"),
  my_measure = "MLN_USD"
)
}
\arguments{
\item{my_measure}{filters the required measure
(US dollars - MLN_USD , US dollars per capita (current PPPs) - USD_CAP).
The default is MLN_USD}
}
\description{
This function imports (nominal) GDP data from OECD format.
}
\details{
This indicator is based on nominal GDP (also called GDP at current prices
or GDP in value) and is available in different measures: US dollars and US
dollars per capita (current PPPs). All OECD countries compile their data
according to the 2008 System of National Accounts (SNA). This indicator is
less suited for comparisons over time, as developments are not only caused
by real growth, but also by changes in prices and PPPs.
}
