% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_OECD_data.R
\name{import.oecd.houseprice}
\alias{import.oecd.houseprice}
\title{Import OECD house price dataframe}
\usage{

  import.oecd.houseprice(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\OECD\\\\House_prices.csv"), my_frequency = "Q",
  my_subject = "NOMINAL")
}
\arguments{
\item{my_frequency}{filters the required frequency
(annual - A or quarter - Q). The default is Q
Import OECD share price dataframe


This function imports share price data from OECD format.}

\item{my_subject}{This parameter filters the required data type
(nominal price, real price, price to rent, price to income or rent).
 The default is real price}

\item{my_frequency}{filters the required frequency
(annual - A , quarter - Q or monthly - M). The default is Q}

\item{my_subject}{filters the required category
(nominal - NOMINAL, real - REAL,
ratio of price to income - PRICEINCOME,
ratio of price to rent - PRICERENT). The default is NOMINAL}
}
\description{
This function imports house price data from OECD format.
}
\details{
The housing prices indicator shows indices of residential property prices
over time. Included are rent prices, real and nominal house prices,
and ratios of price to rent and price to income;
the main elements of housing costs.

In most cases, the nominal house price covers the sale of newly-built and
existing dwellings, following the recommendations from RPPI
(Residential Property Prices Indices) manual.

The real house price is given by the ratio of nominal price
to the consumers’ expenditure deflator in each country,
both seasonally adjusted, from the OECD national accounts database.

The price to income ratio is the nominal house price divided by the nominal
disposable income per head and can be considered as a measure of
affordability.

The price to rent ratio is the nominal house price divided by the rent price
and can be considered as a measure of the profitability of house ownership.

This indicator is an index with base year 2015.

Share price indices are calculated from the prices of common shares of
companies traded on national or foreign stock exchanges. They are usually
determined by the stock exchange, using the closing daily values for the
monthly data, and normally expressed as simple arithmetic averages of the
daily data. A share price index measures how the value of the stocks in the
index is changing, a share return index tells the investor what their
“return” is, meaning how much money they would make as a result of investing
in that basket of shares. A price index measures changes in the market
capitalisation of the basket of shares in the index whereas a return index
adds on to the price index the value of dividend payments, assuming they
are re-invested in the same stocks. Occasionally agencies such as central
banks will compile share indices.
}
