% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.trilemma.ind}
\alias{import.trilemma.ind}
\title{This function imports indexes of financial trilemma (capital openess)}
\usage{
import.trilemma.ind(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\OneDrive - Bank Of Israel\\\\Data",
    "\\\\AizenmanChinnIto\\\\trilemma_indexes_update2018.xlsx")
)
}
\description{
@import readxl
}
\details{
@import dplyr

 @import stringr
}
