% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_WDI_data.R
\name{import.wdi.gdp}
\alias{import.wdi.gdp}
\title{Import World Bank GDP dataframe}
\usage{
import.wdi.gdp(filepath, countries_codes = NULL,
  message_gdp_type = TRUE)
}
\arguments{
\item{countries_codes}{vector of codes of required countries}

\item{message_gdp_type}{boolean}
}
\description{
This function imports GDP data from World Bank format.
}
\details{
Current data are in current U.S. dollars. Dollar figures for GDP are
converted from domestic currencies using single year official exchange rates.
For a few countries where the official exchange rate does not reflect the
rate effectively applied to actual foreign exchange transactions, an
alternative conversion factor is used. World Bank ID is NY.GDP.MKTP.CD

Constant data are in constant 2010 U.S. dollars. Dollar figures for GDP are
converted from domestic currencies using 2010 official exchange rates.
For a few countries where the official exchange rate does not reflect the
rate effectively applied to actual foreign exchange transactions, an
alternative conversion factor is used. World Bank ID is NY.GDP.MKTP.KD
}
