% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.wgi.ind}
\alias{import.wgi.ind}
\title{This helper function imports Worldwide Governance Indicators data}
\usage{
import.wgi.ind(
  filepath = paste0(file.path(Sys.getenv("USERPROFILE"), fsep = "\\\\"),
    "\\\\Documents\\\\Data\\\\World Bank\\\\WGIData.csv"),
  countries_vec = NULL
)
}
\description{
This helper function imports Worldwide Governance Indicators data
}
