% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import_bis_fin_cycle_df}
\alias{import_bis_fin_cycle_df}
\title{This helper function imports credit data from BIS  data base}
\usage{
import_bis_fin_cycle_df(filepath_list = NULL, countries_vec = NULL)
}
\description{
This helper function imports credit data from BIS  data base
}
