% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import_cross_border_balance}
\alias{import_cross_border_balance}
\title{This helper function imports cross border banking from BIS  data base}
\usage{
import_cross_border_balance(
  filepath = NULL,
  countries_vec = NULL,
  annual_freq = TRUE
)
}
\description{
This helper function imports cross border banking from BIS  data base
}
