% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_IMF_data.R
\name{import_imf_df}
\alias{import_imf_df}
\title{This helper function imports data file formatted in IMF Direction of
Trade data base style}
\usage{
import_imf_df(filepath, countries_vec = NULL)
}
\description{
The function returns country pairs formatted in:
small (namewise) country first order
}
