% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import_wdi_df}
\alias{import_wdi_df}
\title{This helper function imports GDP and population data from
WDI  data base}
\usage{
import_wdi_df(filepath_list = NULL, countries_vec = NULL)
}
\description{
This helper function imports GDP and population data from
WDI  data base
}
