% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{make.sens.reg.df}
\alias{make.sens.reg.df}
\title{This function constructs full df for sensitivity regression}
\usage{
make.sens.reg.df(df, bank_int, trade_int, win_len, sector = "GDP")
}
\description{
This function constructs full df for sensitivity regression
}
