% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{make.synch.data}
\alias{make.synch.data}
\title{This function calculates all the synch measures and merges them
into one dataset named accordingly}
\usage{
make.synch.data(df, win_len)
}
\description{
This function calculates all the synch measures and merges them
into one dataset named accordingly
}
