% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{make_strata_reg_list}
\alias{make_strata_reg_list}
\title{This function runs panel regression for each strata separately}
\usage{
make_strata_reg_list(
  reg_df,
  reg_formula,
  my_effect = "twoways",
  my_model = "within",
  criteria_list = NULL
)
}
\description{
This function runs panel regression for each strata separately
}
