% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{normalize.bis.data}
\alias{normalize.bis.data}
\title{This function normalizes the BIS country pairs data
The function divides each country pair position by the
sum of the countries  population (or GDP) in each year
BIS df's structure should be (Date, CountryPair, Balance_Pos, )}
\usage{
normalize.bis.data(bis_df, norm_df, norm_val = "GDP")
}
\description{
This function normalizes the BIS country pairs data
The function divides each country pair position by the
sum of the countries  population (or GDP) in each year
BIS df's structure should be (Date, CountryPair, Balance_Pos, )
}
