% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_functions.R
\name{normalize.imf.data}
\alias{normalize.imf.data}
\title{This function normalizes IMF country pairs data
The function divides each country pair position by the
sum of the countries  population (or GDP) in each year}
\usage{
normalize.imf.data(imf_df, wdi_df, norm_val = "GDP")
}
\description{
This function normalizes IMF country pairs data
The function divides each country pair position by the
sum of the countries  population (or GDP) in each year
}
