% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_WDI_file.R
\name{process.wdi.file}
\alias{process.wdi.file}
\title{Format WDI data}
\usage{
process.wdi.file(temp_file, var_name)
}
\value{
tidy tibble with Country, Year, Varname columns
}
\description{
This function process the data files from
World Development Indicators and converts
the data to tidy format
}
