% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.bis.cpi.data}
\alias{import.bis.cpi.data}
\title{This helper function imports BIS cpi data and converts it to tidy format}
\usage{
import.bis.cpi.data(filepath = paste0("C:\\\\Users\\\\Misha\\\\",
  "Documents\\\\Data\\\\BIS", "\\\\WEBSTATS_LONG_",
  "CPI_DATAFLOW_csv_col.csv"), annual_freq = TRUE)
}
\description{
This helper function imports BIS cpi data and converts it to tidy format
}
